widesight no more crazy, but actually enhances the viewing area by 20% poison is stronger by 100% burning is stronger by 25%

+++ src/system/graphic/Camera.h
@@ -5,7 +5,7 @@
 
 class Camera {
 private:
-	float m_w, m_h, m_halfW, m_halfH;
+	int m_w, m_h, m_halfW, m_halfH;
 public:
 	Camera();
 	~Camera();
